Adding Spiral vb.net code
Adding Spiral vb.net code. This is a translation (interpretation) of the
pseudo-code found in
http://www.sciencedirect.com/science/article/pii/S0167865514001184 into
vb.net application. It is a bit verbose. Utilizing datatables and charts
to demonstrate the process, A more efficient process can focus on the
X,Y of fingerprint templates, the picking of singular point. The
distance calc frrom singular point. The sorting of the poles (those
distances).  Random initial distance key (pole). Then using arctan calc
the right trianle produce. Then raise the poles to form a spiral, as the
sum of the angles. This is similar to spiral of theodorus except there
not all the triangles hypotenus are sequential square root lenghts.